farm_render_deploy/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pandas>=2.1.4
openpyxl>=3.1.2
httpx>=0.25.2
aiosqlite>=0.19.0
python-multipart>=0.0.6